fixed nmap, created setup and some exploits

+++ Spitzer/scanner/scanner.py
@@ -27,35 +27,42 @@ def scan():
         port = config.get_data('ports')
         for p in port:
             ports += str(p) + ','
-        ports = ports[:-1] #remove last comma
+        ports = '-p' + ports[:-1] #remove last comma
     elif scan == 'all':
-        ports = '0-65535'
+        ports = '-p 0-65535'
     elif scan == 'top1':
-        ports = '--top-ports 1000'
+        ports = '--top-ports=1000'
     elif scan == 'top10':
-        ports = '--top-ports 10000'
+        ports = '--top-ports=10000'
     else:
-        ports = scan
+        ports = '-p' + scan
 
     #run masscan x times
     #get results from masscan and create one list of hosts with ports
-    result = {}
-    for i in range(times):
-        print('[-] Starting Masscan ' + str(i+1))
-        masscanner.scan(hosts, ports, rate)
-        xml = chache.read_file('sweep.xml')
-        if xml == '':
-            print_error('Scan '+str(i+1)+' failed!')
-            print_error('Are there any hosts up? Is the interface correct?')
-            return
+    if times != 0:
+        result = {}
+        for i in range(times):
+            print('[-] Starting Masscan ' + str(i+1))
+            masscanner.scan(hosts, ports, rate)
+            xml = chache.read_file('sweep.xml')
+            if xml == '':
+                print_error('Scan '+str(i+1)+' failed!')
+                print_error('Are there any hosts up? Is the interface correct?')
+                return
 
-        print('[-] Masscan '+str(i+1)+' done')
+            print('[-] Masscan '+str(i+1)+' done')
 
-        mass = json.loads(json.dumps(xmltodict.parse(xml, attr_prefix='')))
-        result = host.merge(mass, result)
-        chache.remove_file('sweep.xml')
-        
-    #run nmap once to confirm scan (masscan has some false positives)
-    return nmapper.scan(result)
+            mass = json.loads(json.dumps(xmltodict.parse(xml, attr_prefix='')))
+            result = host.merge(mass, result)
+            chache.remove_file('sweep.xml')
+            
+        #run nmap once to confirm scan (masscan has some false positives)
 
+        print('[*] All masscans are done, found: ')
 
+        for h in result:
+            print(h)
+
+        return nmapper.scan(result)
+    else:
+        return nmapper.run_nmap(hosts, ports)
